requirements.txt
alembic
Authlib
fastapi
celery
redis
httpx
ipython
itsdangerous
Jinja2
psycopg2
pytest
requests
SQLAlchemy
uvicorn
passlib
bcrypt
sqlalchemy-utils
python-multipart
pyjwt